+++ NMSSH/NMSSHSession.m
@@ -248,7 +248,7 @@
 
 - (BOOL)supportsAuthenticationMethod:(NSString *)method {
     char *userauthlist = libssh2_userauth_list(session, [username UTF8String],
-                                               strlen([username UTF8String]));
+                                   (unsigned int)strlen([username UTF8String]));
 
     if (strstr(userauthlist, [method UTF8String]) == NULL) {
         NSLog(@"NMSSH: Authentication by %@ not available for %@", method, host);
